Screen/Linux: Add special handling for AMD DCE-12 display engine.
Same register layout as older DCE's, but the absolute offset of
the DCE IP block inside the BAR has changed, so we need to compute
and apply a corrective absolute offset to the CRTC base addresses
to virtually "shift" the DCE block under DCE-12, allowing us to
otherwise recycle register offsets within the block, crtc offsets,
access methods.

At the moment there is one common ip offset for DCE 12.0/12.1
for Vega 10,12,20. This is not guaranteed to stay that way for
future DCE versions.

Also, it is now confirmed by Alex that DCN is a much redesigned
display engine, with different register layouts and programming,
so DCN Asics are the end of the road for our own MMIO gpu hacks.
Currently the new Raven APU's use DCN-1.0, so for those we need
to find a way to block mmio mapping. Tbd.