XOrgConfCreator: Adapt status output a bit for AMD DCN-1 gpu's.
They get detected as unknown DCE, as they aren't DCE's, and
that maps to assuming the xf86-video-ati ddx being the right
driver. To my surprise that actually works, as the ati ddx
apparently diverts to the proper amdgpu-ddx!

Anyway, adapt status output wrt. this, to aoid confusing
myself and others in the future.