help HybridGraphics: Minimal updates.
PRIME renderoffload now retested on a Intel + NVidia laptop with
Ubuntu 18.04.2-LTS and hwe stack, ie. X-Server 1.20.1 and Linux
4.18. Results:

Linux 4.15 kernel of 18.04 LTS works, as does a tested optional
Linux 5.0 kernel. Both work with modesetting-ddx 1.20.1 of the
server, or with the dedicated intel-ddx/nouveau-ddx as built from
source, as the versions shipping with 18.04.2-LTS hwe are a bit
outdated. Both work with 24 and 30 bit color depth.

What *doesn't* work is the Linux 4.18 hwe kernel! This may be due
to vgaswitcheroo switching failure with nouveau on the very old
2010 MacBookPro Apple laptop, and may not be an issue on other
laptops, or truly muxless laptops without Apple gmux, but this
has not been verified here due to lack of suitable hardware.

So the news are that with server 1.20.1, PRIME + Pageflipping
also works nicely with the modesetting-ddx that is now the
default choice for Ubuntu 18.04.2-LTS HWE, no dedicated intel-ddx
or nouveau-ddx needed or chosen by default. In the past, the intel
ddx was required for PRIME to work properly.

Also color depth 30 works fine with modesetting ddx, but needs
the recompiled head/master git drivers for intel-ddx and nouveau-ddx
as 14.04.2 ships old drivers. This seems to have been rectified at
least for nouveau on Ubuntu 19.04 LTS, but untested so far.

One source of occassional trouble was the gdm3 login manager/greeter
under nouveau on this old NV-A5 GeForce gpu, which occassionally hang
with nouveau-kms error messages in dmesg. Seems to be not very well
written :(. But for most parts it worked.

Ubuntu 18.04.2-LTS was also tested in isolation at depth 30 with
intel, nvidia and amd, so depth 30 seems to be in good shape.

However, note that while NVidia/nouveau and AMD/radeon/amdgpu can
output at 10 bpc / depth 30, the intel-kms drivers still lack proper
gamma hw lut setup code to pass through 10 bpc! They truncate to 8
bpc at the output side. Patches to rectify this are on the mailing
list, but not yet reviewed/tested by myself or afaik others.