Add Danny Garside’s SPTToCCT
Did a little cleanup on the comments in the contributed file, and added to Contents.m.

This routine computes the correlated color temperature of a passed spectrum.  The provided examples check that it gives the right answer on some test cases, and I verified that conversion to 1960 uv matches what I get for that with a different routine (in ISETBio, isetbio.org).  The method here is table lookup.  It seems to perform better than an older analytic approximation (see Wyszecki and Stiles, 1982) for description of that approximation, and code in ISETBio for an implementation.