ScriptingGlue/Python: Make unicode string creation more robust.
The old implementation expected UTF-8 encoded strings, and
could return NULL on invalid UTF-8, e.g., when the string
was in a different encoding.

Now we check for such failures and retry converting the
string assuming it is encoded in the current system locale,
which might be valid for some strings returned by PsychHID
on Windows. On Python 2 we fallback to old style strings,
as it lacks the locale sensitive decoding.

Most importantly, on MS-Windows, after UTF-8 decode failure,
we try use of PyUnicode_DecodeMBCS() as first fallback, to
decode multibyte character strings encoded in the MS-Windows
current system active codepage CP_ACP. The ANSI versions of
Windows api's return strings like HID input device names
in this Windows ANSI codepage and therefore need this decoding,
e.g., for japanese - and presumably chinese - versions of
Windows. Ideally we'd use the Unicode/Widechar versions of
Windows api's that process or return text strings, but changing
this might produce unexpected fallout, so for the moment we
stick to the ANSI api's.

The root cause and this fix for MS-Windows was found by
Hiroyuki Sogo, thanks!

If MBCS decoding fails, we go for decoding according to
current system locale. As Python seems to have no sense of
backwards compatibility, with incompatible changes from
Python 3.0 -> 3.2, then again for 3.2 -> 3.6 and again from
3.6 -> 3.7, sometimes even only affecting one OS, e.g.,
Windows, we implement any half-way meaningful fallback of
trying to decode.

If everything else fails, we assume Latin-1 encoding to either
get a valid Latin-1 string, or something that is essentially
the raw Byte string with values 0 - 255 which may be suitable
for debugging or further workarounds from the calling Python
script.

In any case, don't return NULL, as that leads to crashes,
but at least return some "NULL" or "Invalid" string.

-> This is meant to work around issues found when calling
   PsychHID('Devices', 4) on a japanese MS-Windows 10, which
   does encode device product name strings in locale encoding

   Reference: https://github.com/psychopy/psychopy/issues/2453