Linux: Add detection and workarounds for missing hos XInitThreads().
For multithreading safe operation of PsychHID, e.g., KbQueues, as well
as for Screen(), Xlib must be initialized for multithreaded mode.
This has to happen via a call to XInitThreads() as the *very first*
call to Xlib functions, ie. immediately at application startup and
before any connections to the X-Server are made, error handler are
set etc. ie. before the applications actual GUI starts up.

Calling XInitThreads() later in the application life-cycle can lead to
undefined behaviour ranging from weird bugs to application crash!

On Matlab this was never a problem, as Matlab calls XInitThreads()
properly in all modes of operation, ie. with/without GUI, in
-nodesktop, -nojvm, -nodisplay modes etc.

The same is true for Octave when called as "octave", optionally with
any of the switches --gui, --force-gui, --no-gui, with or without GUI.

Only calling the special stripped down variant "octave-cli" or using
the special command line switch --no-window-system on "octave" would
prevent the correct call to XInitThreads(), but these variants are
never used in practice with Psychtoolbox, therefore there wasn't any
need for PsychHID to deal with this.

The situation is now different with the Python version of PsychHID,
as it turned out that *none* of the tested Python runtime environments
calls XInitThreads() properly! This can lead to occassional crashes.

Try to deal with this by detecting during startup of any X11 enabled
multi-threaded module if XInitThreads() was already called. If so, we
are good and there's nothing to do. If not, try to find out if it is
safe for the module to call XInitThreads() itself, ie. if X-Lib was so
far unused since process creation. If so, call XInitThreads() to fix
multi-threading issues.

For this we implement a Linux-only helper PsychOSNeedXInitThreads()
that checks system setup and makes the decision if a workaround is
or isn't needed and safe. It also provides diagnostic output and
troubleshooting help to the user. The X-Lib using module must call
PsychOSNeedXInitThreads() early during startup before any other X-Lib
calls, and call XInitThreads() itself if advised by the helper function.

Currently only PsychHID() and Screen() need this treatment and now
use this method of deciding on the workaround.

This approach seems to work with the following so far tested python
applications on Ubuntu 18.04.2 LTS: python and python3, launched
without a GUI from the terminal. Also the Python IDE's "eric" and
"thommy".

What doesn't work is the PsychoPy IDE as of version 3.1.2. It doesn't
call XInitThreads() at all, but has made many XLib calls at the time
PsychHID gets loaded, thereby preventing automatic workarounds.

This app needs to be fixed at the source, possibly by using PyQt very
early during application startup, as at least Octave's QT-5 based
GUI does proper threading setup during QT init. Or better by
importing XLib bindings and calling XInitThreads() directly as
the first thing during application startup if running under X11.

As this will be fixed properly in upcoming PsychoPy 3.1.3, a hint
to the user is presented on trouble, advising an upgrade to PsychoPy
3.1.3 or later.

Technical Note:

There isn't any official api to detect if XInitThreads() was called,
or if XLib was so far idle. Therefore we use the following hack, which
hopefully will hold up on systems other than the tested Ubuntu
18.04.2 LTS: Use dlsym() to get references to libX11 internal global
variables and then check those variables values against expectations.
_Xglobal_lock should be NULL if XInitThreads() wasn't called yet,
non-NULL otherwise. _XErrorFunction should be NULL immediately after
libX11 loading and before X11 calls are made. Calls to XSetErrorHandler()
as part of setting up XLib for use, or to XOpenDisplay() as part of
a GUI startup sequemce, will assign a proper error handler to
_XErrorFunction, so it becomes non-NULL - a sign for us to not call
XInitThreads() but instead warn the user about pending doom. The hack
assumes use of a libX11 as shipped by XOrg, as i verified its internal
working and compatibility against libX11 master.