DownloadPsychtoolbox: Give hint about outdated svn clients under Matlab.
We can't do this on Octave due to limitations of its system()
function. It can either give valuable live output for feedback
on svn progress, or return the svn output in a variable, but not
both.

On Matlab we can have it both ways, so retrieve and parse the
svn output result string on svn failures and give user a hint
about outdated svn clients if the status message of svn hints
at that reason for download failure.

-> Mostly untested, apart from basic compatibility, as i don't have
   the kind of broken Apple macOS systems which would trigger this
   error handling.