setup.py: Fix line-endings, PsychPortAudio on 32-Bit Linux.
Line endings were switched to CR+LF MS-DOS style, but we use
Unix LF style. Revert that.

PsychPortAudio can be built for 32-Bit Linux machine
architectures, as modern Linux ships with the required
and dynamically linked libportaudio. This matters, e.g.,
for building on embedded/mobils SoC's like the
RaspberryPi with 32-Bit ARM architecture. On macOS it
doesn't matter as it is 64-Bit only. Only 32-Bit Windows
on Intel is a no-go, because we don't ship the neccessary
compiled 32-Bit dll.

Add a comment to the -mmacosx-version-min=10.9 switch that
a setting lower than 10.11 is not tested or supported by
Psychtoolbox upstream in case of problems. Given that Apple
cancelled all support for 10.11 already, even running a
supported 10.11 system nowadays is a computer security risk
and 10.13 would be a more sane setting.