Fix buglet from PR
The PR from MS correctly removed hard coding of S, which should be and is passed.  But it didn’t fix the reference made to that variable, which should have been changed to S_SPD.  That is now done.  This was easily caught when I executed the examples that are in the source code.