VRRTest: Add support for hw timestamping via VideoSwitcher/RTBox/Bits#
The VideoSwitcher + RTBox combo was tested using the VideoSwitcher
plus the emulated/built-in RTBox of the Bits#. Doesn't work fully
reliable. My suspicion is that the VideoSwitcher can't emit a 5V
TTL pulse in all trials, and occassionally only emits a lower voltage
pulse, so the Bits# doesn't detect the trigger and stuff gets out of
whack. This is very plausible, and wouldn't be the VideoSwitchers fault,
because i lack a compatible 5V power-supply for German power sockets,
so have to use the switcher without external power. This miraculously
seems to work most of the time, just sucking power from the VGA signal,
but probably makes the trigger pulse voltage unreliable. Would be good
to retest this with a proper power supply connected.

The Bits# method uses a T-Lock code to emit a BNC trigger signal and
then a BNC loopback capble to feed that into the BNC trigger input
port and hook it up to the trigger input of the emulated RTBox, for
stimulus onset timestamping. Now the approach shows quite some flaws or
bugs of the CRS Bits# design and RTBox emulation, and the misdesigned
T-Lock mechanism, that doesn't emit triggers in the frame where the
T-Lock is, but in the *following* scanout cycle (!?!), makes this
method of timestamping rather fragile in case of repeated stimulus
frames due to missed flip deadline ("skipped frame") or VBLANK timeout
in VRR mode. In such cases, which get triggered whenever BTR/LFC
low-framerate compensation gets engaged, timestamps from the Bits#
will be utterly bogus. At least for measurements well within the
safe MinVRR - MaxVRR FreeSync range, this allows somewhat reliable
validation of FreeSync timestamping against hw measured ground truth.

Testing, so far on a Raven Ridge AMD Ryzen 5 2400G with Vega11
integrated gpu and DCN-1.0 displays engine under upcoming Linux 5.3,
shows timestamping of Screen('Flip') in VRR mode to be correct, within
the testable range of the Bits# method. One oddity is that the
'Flip' stimulus onset timestamps seem to be about 1 msec delayed
wrt. ground truth at a 85 Hz video mode and VRR range 33-85 Hz.
Not sure yet what causes this discrepancy, but stddev of the error
is very small and the small constant offset means that timestamping
of the OS + Psychtoolbox is essentially working well.