PsychtoolboxKernelDriver: Update Xcode configuration.
This makes the driver build again. I think i deleted the PPC version of
the
user client tool, because Xcode doesn’t like that anymore on OSX
10.13.6.

I may have done other things do deal with this piece of crap of a build
tool. Anyway it works for the moment.