ScriptingGlue/Eyelink: On Matlab use mexCALLMatlabWitTrap() for R2018a+ compatibility.
mexSetTrapFlag() was removed in R2018a, so to build our mex files
with R2018a+ we need to replace mexSetTrapFlag() + mexCallMATLAB() by
an equivalent call to mexCallMATLABWithTrap().

A new helper function wraps such calls, doing it the new way
for Matlab, and the old way for Octave, as Octave doesn't
support mexCallMATLABWithTrap() in version 4.0.0. Octave 4.2
does support it, but we still need to be able to build and test
with Octave 4.0 for Ubuntu 16.04.6 LTS compatibility.

-> Mex files build successfully against R2019a on Ubuntu 19.04
   and against Octave 4.4.1.

-> Note that these mex files built on Ubuntu 19.04 won't run on
   18.04.2-LTS, because the libm math library seems to request
   some GLIBC_2_29 symbols, where 18.04-LTS only supports
   GLIBC_2_27. Iow. 19.04 can't substitute as build system that
   allows backwards compat with 18.04-LTS, at least not without
   more work.