PsychLinuxConfiguration: Fix for gamemode.ini setup for compat with Matlab.
Turns out that in case of an empty XDG_CONFIG_HOME (which is default
on most distros), a file '/gamemode.ini' is searched by exist().

On Octave this rightfully reported no such file.

On Matlab however, this finds a match because the file
is in Matlab's Psychtoolobx search path, duh!

Fix logic, so initial install of gamemode.ini actually works
not only on Octave but also Matlab.

Problem and almost correct solution reported by iandol
in GitHub issue #145.

Fixes #145.

-> Needs pick into 3.0.15 "final".