Screen/Windows: Kill vblank timestamping. It's broken on Win-10
Turns out that PsychOSGetVBLTimeAndCount() is utterly broken at
least on Windows-10. Now PsychDwmGetCompositionTimingInfo()
returns timing info with success status code even if the compositor
is out of the way on a single display setup in fullscreen mode, ie.
when the single fullscreen onscreen window is unredirected.

Of course the results are utter garbage, and cause Screen
'Getwindowinfo' to return wrong results for vblank count and
vblank time. That causes PredictVisualOnsetForTime() to
mispredict, which, e.g., causes PsychPortAudioTimingTest to
screw up audio-video sync. Yay!

According to commit messages from 2012 this supposedly
worked one time on some Windows-7 or Windows-8 release,
but no more on at least my Windows-10 single display setup.

Give up on this, it is for the better.