PsychOculusVRCore/Windows: Update for build against R2019a + MSVC 2019
For building with R2019a and MSVC 2019 we had to hack the OVR_CAPI.h file to #define OVR_ALIGN(n) as nothing. Hopefully this still
works, as i don't have a Rift DK2 around for testing atm., and my
Windows 10 box refuses installation of the full runtime drive anyway.

For the Matlab build, we compile the OVR_CAPIshim.c shim file and
other helper C files from the SDK directly into our mex file,
instead of statically linking against LibOVR.lib, as that .lib
import file only works with up to MSVC 2013, but not with our
MSVC 2019 build system for R2019a. The shim will locate and
runtime-link against the libOVRRT_64_0_5.dll of the installed
Oculus VR runtime during initialization. Or so i hope.

The build for Octave 5.1.0 did not need any modifications or hacks.

-> OculusVR SDK 0.5.0 is no longer available for download from
Oculus, not even in the legacy section. I have a backup copy here
locally of the SDK and runtime. This will pose a problem for anybody
who would want to use this driver on Windows.

-> Clearly this driver is on its last leg for Windows, maybe we should
just let it  r.i.p.?