PsychHID/Linux: Fix race condition in KbQueueStop->KbQueueStart cycles.
Before enabling X-Event dispatch to keyboard/touch queues and starting
the processing thread, drain our X event queue from any potential stale
events from previous runs.

This tries to address a potential race condition that could happen during a
previous PsychHIDOSKbQueueStop() cycle, when the X-Server would enqueue new X-Events
while we are already in shutdown, so the KbQueueThread would terminate before it can
dequeue and discard those fresh events in the X event queue. User action / input,
KbQueueProcessEvents() execution and call of KbQueueStop by user script would have
to happen in just the right order within a small (< 1 msecs?) time window for this
spillover to happen, but it is not impossible. This will hopefully help against the
spillover problem reported in forum message #23899 when using a touch screen.

I could not reproduce this problem, but i hope this is what happened
on forum message 23899, so this should fix it.