GetSecs: Add 4th return argument monotonicTime for 'AllClocks'
Returns the system monotonic timebase, which is == GetSecs for
Windows (usually QueryPerformanceCounter()) and macOS
(mach_absolute_time()), but different on Linux, where it is
clock_gettime(CLOCK_MONOTONIC,...); for the Posix montonic
clock.

This is mostly useful on Linux for syncing clocks with 3rd party
libs like the Tobii eyetracker runtime.