Screen/OSX: Make timing fix Retina compatible.
Since pageflipping is completely broken under the Cocoa
backend with macOS 10.14.6 Mojave and 10.15.0, we have
to use CGL for both non-Retina and Retina displays.

As CGL does not support Retina scaling in any way, and
CGL backbuffer scaling usually kills pageflipping as well,
we use the following hack:

We try to find the native display mode of a builtin/Retina
display and then modeswitch to it. Then use that to drive
the Retina display directly at native resolution for page-
flipping under CGL. Restore mode at end of session.
This is dumb, but at least it works under 10.14.6 Mojave,
as the only known way to fix Retina display timing.

Also restructure matching logic, improve diagnostic
debug output etc.

So far tested on 10.14.6 with MBP 2017 AMD Polaris,
with external standard DPI display - a VGA monitor -
and on the internal laptop Retina panel.