Screen/OSX: Move screen->head auto-detection to safer place.
We used to call PsychAutoDetectScreenToHeadMappings()
at Screen() init time, but now, just as on Linux, we call it
at the end of PsychOSOpenOnscreenWindow(), after the
OpenGL context for the window has been created.

Why? Because at least on modern MacBookPro's, e.g., the
2017 15 inch MBP with AMD Polaris gpu, the discrete gpu
from AMD may be powered down before the OpenGL
context is created. Accessing the powered down gpu for
register reads in PsychAutoDetectScreenToHeadMappings
can cause a hard fault, leading to a unhandled machine
check exception (MCA/MCE CATERR) and therefore to a
hard crash -> emergency reboot --> Not what we want as
user experience.

This new way of calling after OpenGL context creation,
we can be sure that the AMD gpu is actually powered up
and online when we access it first.

-> Fixes hard crashes of MBP 2017 if automatic graphics
switching is active.