Screen/OSX: Deal with Apple iMac brain-damage via lookup table.
The Apple iMac's internal flat panels, despite having a native
resolution, are not exposed by the trainwreck as having a native
preferred optimal resolution. Instead they pretend to be like
CRT monitors that can operate at different video resolutions with
proper timing etc. Apparently the gpu's hw scaler is used to
rescale from the PTB selected framebuffer size aka "video mode"
to the true native panel size, iow. the gpu frontend and backend
resolutions are mismatched. This doesn't seem to affect pageflipping
and visual onset timing, but it does screw up our beamposition
timestamping, as we need the true vbl startline for that, not the
fake framebuffer height / video mode height.

So where do we get it from? Various approaches were considered,
e.g., direct vblank start position readout from the gpu, but all
had drawbacks like high dependency on the PTB kernel driver, and/
or higher maintenance overhead for the driver, or ugly heuristics
that may go wrong, given that we have to rely on macOS broken and
lying video mode api's.

In the end, the least awful solution seems to be a lookup table
that looks up the true panel native pixel size by Mac modelId and
panel width in mm and uses that for beamposition setup. This will
require maintenance for all future Mac's that have builtin panels
and this brain-damage implemented, but at least maintenance by
editing a LUT and recompiling Screen() is easier than maintenance
of the kernel driver. So far, only iMac's seem to suffer from this
special type of misdesign, the Apple MacBook line laptops and external
displays seem to be not affected by this stupidity. This keeps the
effort low enough to only dealing with new iMac models.

-> Survived basic test on a MBP 2010, test on actual iMac pending.