Screen: Improve gpu detection, especially on Linux.
In addition to checking for GL_RENDERER ATI or AMD, also check
for Radeon, as some Radeon's on at least Linux no longer have
AMD or ATI in their renderer name, and vendor X.Org is not that
descriptive on a AMD gpu.

Moreover, on Linux/X11 with Mesa, use the Mesa_query_renderer
extension to find the pci vendor id of the gpu vendor of the
render gpu, as that is much more bulletproof than any name
matching heuristic, and should work at least on modern distros
with PCI(e) gpu's. We currently match for AMD, Intel and NVidia.