Screen: Implement basic infrastructure for VRR on AMD FreeSync, NVidia G-Sync.
This cleans up code related to VRR, adds basic detection if VRR
is (likely) active via a new helper PsychVRRActivce(windowRecord).

It adds detection code to find out if we are running on a single-gpu
setup with a G-Sync capable NVidia gpu under the proprietary driver,
ie. if G-Sync mode is possible on the given setup.

It adds workarounds for NVidia's G-Sync implementation as provided
by the proprietary NVidia driver, ie. to make the startup test wrt.
beamposition queries and sync not fail due to the peculiar behaviour
of beampos queries and timing in G-Sync mode. And to make beamposition
timestamping work again in such a G-Sync configuration.

The basic insight here is that the scanout position register in the
GPU will stop/freeze at the last line of the fixed minimum duration
front-porch iff the GPU is in G-Sync mode and no pageflip is pending
in the display hardware at time of reaching min front-porch. Iow.,
the counter is stopped while inside the variable length / duration
extended front-porch. After front-porch ends due to timeout or due
to flip programmed, executed and finished, and the display engine
enters back-porch, the counter will start running again as usual,
counting time until end of vblank / start of active scanout etc.,
just as in fixed refresh rate mode. So what we do is we don't check
for stuck beamposition queries during initial startup on NVidia
Kepler and later (as Kepler and later are G-Sync capable), and don't
check for stuck beamposition if G-Sync mode is active. During beampos
calibration of vbl endline and refresh, we issue a bufferswap request
during each refresh cycle to keep the counter "unstuck", avoiding
bogus results and or 1 second stalls for each cycle until timeout.

At runtime we "beamposition timestamp" as usual: Each flip completion
means that a flip was pending and completed in the vblank/scanout
of completion, ergo the scanout counter can't be stuck, ergo it
can be used to measure scanlines (and thereby time) until end of
back-porch / vblank or elapsed time since end of vblank / start of
active scanout. The only catch is that VBLTimestamp is meaningless,
we can only determine StimulusOnsetTime, because we don't know how
much time has elapsed during extended variable duration front-porch.
Therefore we set VBLTimestamp = StimulusOnsetTime as the closest
approximation we can make -- wrong, but still somewhat useful for
flip scheduling in G-Sync mode.

Note that this may still need future adjustments, it is a 1st working
version, as verified on a NVidia GeForce 1070GTX single display G-Sync
setup with 435 series proprietary drivers with an ASUS Rog G-Sync
panel. Testing with a photometer abused as photo-diode (ColorCal2 +
RtBox) to capture light flashes from VRRTest.m showed returned
Flip timestamps to be at least approximately right wrt. measurement,
and reasonably stable/jitter free. However, this measurement method
is far from robust and not exactly sub-millisecond accurate. The best
i could confirm is that the result is within a 2 milliseconds ballpark
of what i'd expect.

-> Linux implementation only. Unclear if it could be translated to
   Windows.

Next step: Enable VRR/FreeSync support for the AMD side of things.