Screen/Linux: Disable 'software sync' home-made VRR support for AMD.
Disable the low-level implementation for legacy AMD gpu's which
don't have native VRR support.

This was a neat idea that unfortunatly never worked reliably on any
tested digital display (VRR capable or not), and sadly not at all on
any tested analog VGA CRT monitor at which i was aiming this.

We don't delete the PsychOSScheduleSoftSyncFlip() implementation yet,
because some of its ideas might be actually useful for VRR as a
reference.