Add PsychPhotodiode() for photo-diode based visual onset timestamping.
This new function allows to (ab)use a sound-card via PsychPortAudio
for visual onset timestamping. The idea is to connect a photo-diode
or such via suitable amplifier to the microphone/line-input of a
sound card, then use PsychPortAudio to capture with timestamps from
the soundcard, threshold signal spikes and timestamp accordingly
for cheap and easy spot-checking of visual onset timing, comparable
to Screen('Flip')'s timestamps.

First user of PsychPhotodiode() is VRRTest.m, which now has a new
hwmeasurement method 5, using this. Verified to work reasonably
with my photo-diodes and an external USB soundcard on the Retina
panel of my MBP 2017 "groovy" under Ubuntu Linux 19.10. Also minor
cosmetics to VRRTest.m.