PsychPhotodiode: Add 'lrMode' handling, minor cleanups.
'lrMode' to control if on a multi-channel stereo sound signal,
the 1st channel, 2nd channel, or a sum or an average of both
channels should be used for the triggering and stuff.

Default is lrMode 0 = Sum of stereo channels, if any.

This is only well tested for 0 = Sum and 1 = Left channel so far.

Some results wrt. accuracy of the photo-diode method:

CRT monitor: stddev of jitter 13 usecs, < 0.2 msecs error.
MBP 2017 Retina panel: < 2.5 msecs error on good days.
MBP 2010 panel:        < 4 msecs error on good days.

-> On panels one should crank display brightness to the max, to
   avoid or minimize backlight pwm modulation artifacts.

-> This is in a reasonable shape for use with displays where one
   can't use VPixxx devices, VGA VideoSwitcher + RTBox etc.