Screen/Linux: Fix monitor detection for VRR / Freesync.
Due to a thinko, if the Freesync monitor was attached to a video
output other than the first one, Screen() would not detect it
VRR capable, as it always only tested output 0, which is not
always our primary output! This worked by accident on my old
test systems, but not on the actual machine with the actual
FreeSync monitor, which is attached to output 2 :/

Improve the mapping logic, so we will always probe the primary
output of a X-Screen if that output is connected and active,
otherwise probe the first active connected output, otherwise
probe the first output. Just as we do everywhere else when
outputs matter.

-> Now works on darlene with the Freesync nonitor at the 3rd
   output DisplayPort-1.