XOrgConfCreator: Allow users to opt-out of modesetting-ddx.
Ubuntu 20.04 LTS will use modesetting-ddx by default on Intel
and NVidia+nouveau, so make sure users can switch to nouveau-ddx
or intel-ddx if they answer 'n'o to the question if modesetting
ddx should be used.

Essentially we have inverted defaults now: Instead of modesetting
being the opt-in, with intel/nouveau being the default, we now
have modesetting-ddx the default, with nouveau/intel being the
opt-ins.