XOrgConfCreator: Make it robust against different color depths.
Especially on a slightly botched config, where the X-Screen
runs at depth 30, but Mesa doesn't expose 30 bit due to
configuration error, we want XOrgConfCreator to be resilient
against this, because it is part of the potential config fixup
process.

So force 24 bit color depth windows for gpu detection, as those
should always work.