"help PsychContributed": Add a warranty and support disclaimer.
Make it clear that code in this folder does not play by our usual
rules wrt. backwards compatibility, quality and support, and that
users are completely on their own in case of trouble. That we may
even remove functionality in the folder without any warning.

When we're at it, add some reference to Denis new soon to be merged
MacDisplaySettings() function.