MacDisplaySettings.applescript and .m improved speed and error handling.
May 15, 2020. Now also pass a flag from MacDisplaySettings.m to MacDisplaySettings.applescript indicating whether Psychtoolbox has a window on the main screen. In that case, AppleScript will not try to show a dialog. Added a loop in AppleScript to wait for System Preferences window to open; this fixes a rare error.

May 15, 2020. APPLESCRIPT: Added a loop in AppleScript to wait for System Preferences window to open; this fixes a rare error. Replaced every error code with a message in errorMsg.

May 17, 2020. Improved handling of empty args, which are now replaced by default values, just like missing args.

May 20, 2020. APPLESCRIPT: MacDisplaySettings hung up on my student Benji Luo with the Night Shift panel showing. I suspect it was in an endless loop waiting for the menu to pop up after a click. I rewrote the loop to throw an error if the menu doesn't appear after three attempts of clicking and waiting up to 500 ms each time.