KbCheck(): Only use PsychHID('KbCheck') on Windows if checking a non-keyboard.
Iow. if deviceNumber refers to a mouse-type or joystick-type device. For
keyboard or keypads (which are indistinguishable by DirectInput, so just
check for kbs) PsychHID does not have any benefit, given that individual
keyboards can't be individually selected on Windows. Otoh. the hacks needed
to make PsychHID('KbCheck') work on Windows with DirectInput do interact with
KbQueues in surprising ways, so avoid this path unless it brings real benefit
- like the ability to use mouse/trackpad/joystick/gamepad buttons as "keyboard"
response buttons.

See https://psychtoolbox.discourse.group/t/psychhid-kbcheck-starts-event-queue-under-windows-10/3223/6
for the problem. A more elegant fix would require implementing more complex
state tracking between KbCheck and Kbqueues and meticuluous testing, for which
there ain't not time at the moment, so this little band-aid has to do for now.