Scripting glue Matlab: Actually make int64 and uint64 argument types work.
The probing -> mapping from Matlab types to internal constants was
missing! This made PsychCopyInPointerArg() fail in all cases, as it
expects a uint64() value for encoding a 64-Bit pointer on 64-Bit
systems.

This was not ever noticed since none of our existing mex files does
actually use PsychCopyInPointerArg(). Soon a new mex file will, so
fix this.