Screen: Allow use of beamposition vblank timestamping in PsychOSGetVBLTimeAndCount()
If the OS specific timestamping mechanism is unavailable - or broken
to the point of being utterly useless as on macOS and Windows, then
allow use of beamposition timestamping (if available on the current
configuration) to compute vblank timestamp. We use the end-of-vblank/
start of active scanout timestamp.

This should become useful if 3rd party code wants to use
Screen('GetWindowInfo', window, ...) to get a precise vblank
timestamp of the most recent (or ongoing) VBLANK.