Screen: 'GetWindowInfo' now also returns new field 'GLDeviceUUID'.
'GLDeviceUUID' returns the 16 Byte universally unique identifier of
the graphics device (gpu) which is used by the OpenGL context of the
given onscreen window for rendering. This UUID is not quite as
universally unique as the name suggests, but it should be unique at
least for gpu's in a system that are driven by the same OpenGL driver.

This only works if the EXT_memory_objects extension is supported,
which is currently not the case by many drivers.

Example for UUID's:

Linux AMD Mesa radeonsi for GCN gpu's: The UUID is built of 4 groups
of 4 bytes each and encodes the location of the gpu on the PCI bus:

pcidomain-busnumber-devicenumber-devicefunction

The same format is used by the Mesa radv Vulkan driver for AMD.
The AMD provided open-source Vulkan driver amdvlk uses this format
instead as of 2020-Q2-4:

busnumber-devicenumber-devicefunction-0 <- 4 Bytes of trailing zeros.


NVidia proprietary on Linux: Something that looks like a 16 Byte
hash of whatever? Maybe also PCI location, maybe some builtin UUID,
in any way supposed to be unique enough.

I assume the NVidia and AMD drivers on Windows-10 will use similar
style to the Linux variants.

Intel gpu's on both Linux and Windows do not support the required
extensions and therfore likely won't support UUID's yet.