Screen/Linux: Allow more flexibility for Screen('ConfigureDisplay','Scanout').
Allow the theoretical option to also query 'Scanout' state for disabled
outputs. This is very half-baked and needs more work, so for now a
new bool parameter was added, but all callers call it with a FALSE
setting, which gives exactly the same behaviour as in the past.

Setting new 'Scanout' settings of newwidth=newheight=hz == 0 now allows
to disable a video output. The catch: We can't reenable it again after
programmatically disabling it, because 'Scanout' can only deal with
active outputs atm., so it won't be able to access a formerly disabled
output to reenable/modeset it again, yay! This needs way more thought
than i have time for atm. to implement this properly.