PsychVulkanCore: Add new function 'GetHDRProperties' to query HDR display properties.
hdr = PsychVulkanCore('GetHDRProperties', vulkanWindow); queries the HDR display
properties of the display on which 'vulkanWindow' is displaying, and returns
a struct 'hdr' with the data. Mostly if the  query of HDR properties was
successfull or not (e.g., if the display is not HDR capable or the query
failed), and then what are essentially the mastering display HDR properties,
as reported by the displays EDID, e.g., min/max/average luminance in nits,
and color gamut ie. color primaries and white-point.