PsychVulkanCore/Windows: Add HDR property workaround for AMD.
As of AMD Radeon Software 2020.4.2 (June 2020), the AMD Vulkan driver
for Windows-10 still is buggy wrt. reporting maxLuminance as maxFall, and
not reporting maxFall. Apparently nobody was motivated to do what is
presumably a 1-line copy & paste fix...

We give up for the time being and use the DXGI HDR display property
query fallback like on NVidia if maxFall is reported as zero on the AMD
proprietary driver for Windows.