Screen/Linux: Fixup vrr RandR output query commit.
Commit 1c93d01a4e44aced41eaa77721319c60b78bb581 introduced
a little error in that it may not properly disable VRR mode
if VRR mode shall not be used/disabled. As vrr_supported=FALSE
now whenever vrr_wanted is FALSE, we skipped some code that
must run whenever vrr_supported should be true but isn't anymore,
iow. code to actively disable VRR if it is supported by not
wanted.