MacDisplaySettings: Better error handling.
Improved error handling. More argument checking.
Error messages are more helpful. More wait loops
in applescript so as not to fail before the object
becomes available, with limited number of iterations,
to avoid infinite waits. Pass new flag to applescript
routine that indicates whether Screen has a window
open on the main screen, to help applescript do the
right thing if there is an error. (If there is an
error and Screen has a window on main screen, then
we return to MacDisplaySettings.m which calls "sca"
before flagging the error.)